Fixing lick frame off-by-one error
it used to be the case that the lick sensor was polled on every frame in the stimulus code, giving a 1:1 correspondence between the frame number and the index in the 'lickData' array. However, that assumption was violated when we began displaying a frame at the beginning of the session without a corresponding call the 'checkLickSensor' method. Using the 'responselog' instead will provide a more accurate measure of actual like frames and times.