Looks in the correct place for the rig_id
The rig_id is being supplied by MPE in a top-level
field called 'rig_id', not in the 'platform_info'.
The 'platform_info' does also include a field 'rig_id',
but it is not currently saving the right information (it
just says 'unknown'. This commit also changes when we
fall back to using devices.get_rig_id(). Any old data
will simply not have the 'rig_id' key, so if we get a
KeyError we should try to look it up using the computer
name.