Linting and test fixes
A test was setting the date for "new data" to devices.VALID_BEFORE_DATE+1,
which failed with my temporary datetime.max placeholder there. However,
What we want to have happen is for the date to invalidate the mapping for
any experiment on or after the switch, so I changed the test to reflect this.
Now, we assert that an experiment on the same day as the switch will log an
error and not attempt to use the mapping.