Fix/issue509 (#512)
* Get rig ID from data instead of from hardcoded mapping in devices

Since the rig ID is now included in the data, we should read it from there
instead of using the mapping in devices to get the rig ID from the computer
name. This commit does the following:

* Adds a new function to translator/foraging2/extract called get_rig_id().
  This new function looks in the data file for the rig ID. Since this property
  will either not exist or be 'unknown' for old data files, in either of those
  cases we will use the mapping in devices to try and get the rig name. There
  is an additional check to make sure that the experiment date is from a time
  when the mapping in devices is valid.

* foraging2.data_to_metadata() now uses extract to get the rig ID instead of
  using devices.

* devices now includes a constant called VALID_BEFORE_DATE, specifying the
  date before which the mappings in that file are valid. We need to keep this
  mapping around so that we can use it when we load old data, but when MPE
  changes the computers this mapping will break, so we want to be explicit about
  when the mapping was valid.

* Session metrics should not get rig ID from devices

This commit causes session_metrics.rig_id() to use the 'rig_id' from the
trials dataframe instead of automatically trying to use the mapping in devices
to get the rig ID based on the computer name.

In tests/change_detection/trials/test_summarize there is a test called
test_session_level_summary which wants the output rig_id field to be 'unknown'
(see session_summary fixture, tests/conftest.py, line 258) for an input trials
dataframe that does not have a 'rig_id' key (see mock_trials_fixture,
tests/conftest.py, line 193). This was the motivation for including the
KeyError exception handling in session_metrics.rig_id().

* Add tests for translator.foraging2.extract.get_rig_id()

Adds tests for 3 situations:
* New data with the rig_id included (don't fall back to devices)
* Old data without the rig_id (do fall back to devices)
* New data that doesn't have rig_id (don't fall back to devices, warn the user)

* Minor formatting fixes to keep the linter happy

* Looks in the correct place for the rig_id

The rig_id is being supplied by MPE in a top-level
field called 'rig_id', not in the 'platform_info'.
The 'platform_info' does also include a field 'rig_id',
but it is not currently saving the right information (it
just says 'unknown'. This commit also changes when we
fall back to using devices.get_rig_id(). Any old data
will simply not have the 'rig_id' key, so if we get a
KeyError we should try to look it up using the computer
name.

* Updated tests for rig_id extraction

After I got the correct location of the rig_id from MPE and
changed the way the extractor works, I had to change the test
code to reflect the right place for extractor to look.

* Forgot to include comp name in test data structs

* Set the VALID_BEFORE_DATE in devices to year 9999 for now

* Linting and test fixes

A test was setting the date for "new data" to devices.VALID_BEFORE_DATE+1,
which failed with my temporary datetime.max placeholder there. However,
What we want to have happen is for the date to invalidate the mapping for
any experiment on or after the switch, so I changed the test to reflect this.
Now, we assert that an experiment on the same day as the switch will log an
error and not attempt to use the mapping.

* Set VALID date in mapping to 2019-05-01

* Clean up devices todo comments

* The change from using rig_id to comp_id

The rig_id field currently points to the cluster, not to the individual rig.
We want to use the information included in comp_id for now, until MPE changes
the information stored in rig_id.

* Add comp_id to unit test fixture.

This allows us to use the comp_id field for now, until MPE fixes
rig_id.