Use un-rebased times in validate_reward_when_lick_in_window
Addresses issue #422 by making this validation function use the trials
from core_data instead of the extended trials dataframe. During creation
of the extended trials dataframe, the event timestamps are assigned to the
time of the next frame. This behavior interferes with this validation
function, which needs to check that the camstim logic was correctly
executed by comparing the original event timestamps.